requirements.txt
asgiref==3.7.2
Django==5.0
sqlparse==0.4.4
typing_extensions==4.8.0
psycopg2
Pillow~=10.1.0
djangorestframework~=3.14.0